second commit

+++ src/main/java/java12/dao/impl/HouseDaoImpl.java
@@ -2,11 +2,10 @@ package java12.dao.impl;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
+import jakarta.persistence.Query;
 import java12.config.Config;
 import java12.dao.HouseDao;
-import java12.entity.Customer;
-import java12.entity.House;
-import java12.entity.Owner;
+import java12.entity.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,14 +15,72 @@ public class HouseDaoImpl implements HouseDao {
 
     @Override
     public String save(House house, Long ownerId) {
+        EntityManager entityManager = entityManagerFactory.createEntityManager();
+        Owner owner = null;
+        RentInfo rentInfo = new RentInfo();
+        try {
+            entityManager.getTransaction().begin();
 
-        return null;
+            owner = entityManager.find(Owner.class, ownerId);
+            if (owner != null) {
+                entityManager.persist(house);
+                house.setOwner(owner);
+                house.setRentInfo(rentInfo);
+                rentInfo.setOwner(owner);
+                rentInfo.setHouse(house);
+
+                entityManager.persist(house);
+                entityManager.persist(rentInfo);
+            }else {
+                return "Not found owner!";
+            }
+            entityManager.getTransaction().commit();
+        } catch (Exception e) {
+            if (entityManager.getTransaction().isActive()) entityManager.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        }finally {
+            entityManager.close();
+        }
+        return  "saved" + house.getHouseType();
+    }
+
+    @Override
+    public String saveHouse(House house) {
+        EntityManager entityManager = entityManagerFactory.createEntityManager();
+        Owner owner = null;
+        RentInfo rentInfo = new RentInfo();
+        try {
+            entityManager.getTransaction().begin();
+
+                entityManager.persist(house);
+            entityManager.getTransaction().commit();
+        } catch (Exception e) {
+            if (entityManager.getTransaction().isActive()) entityManager.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        }finally {
+            entityManager.close();
+        }
+        return  "saved" + house.getHouseType();
     }
 
     @Override
     public House getById(Long id) {
+        EntityManager entityManager = entityManagerFactory.createEntityManager();
+        House house = null;
 
-        return null;
+        try {
+            entityManager.getTransaction().begin();
+
+            house = entityManager.find(House.class, id);
+
+            entityManager.getTransaction().commit();
+        } catch (Exception e) {
+            if (entityManager.getTransaction().isActive()) entityManager.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        }finally {
+            entityManager.close();
+        }
+        return  house;
     }
 
     @Override
@@ -34,16 +91,32 @@ public class HouseDaoImpl implements HouseDao {
             entityManager.getTransaction().begin();
 
             house = entityManager.find(House.class, id);
-            entityManager.remove(house);
+
+            if (house != null) {
+                Query rentInfoQuery = entityManager.createQuery("select r from RentInfo r where r.house.id = :houseId");
+                rentInfoQuery.setParameter("houseId", id);
+                List<RentInfo> rentInfoList = rentInfoQuery.getResultList();
+
+                if (!rentInfoList.isEmpty()) {
+                    for (RentInfo rentInfo : rentInfoList) {
+                        entityManager.remove(rentInfo);
+                    }
+                }
+                entityManager.remove(house);
+                System.out.println("Deleted " + house.getHouseType());
+            } else {
+                System.out.println("House with ID " + id + " not found.");
+            }
 
             entityManager.getTransaction().commit();
         } catch (Exception e) {
-            if (entityManager.getTransaction().isActive()) entityManager.getTransaction().rollback();
+            if (entityManager.getTransaction().isActive()) {
+                entityManager.getTransaction().rollback();
+            }
             System.out.println(e.getMessage());
-        }finally {
+        } finally {
             entityManager.close();
         }
-        System.out.println("deleted " + house.getHouse());
     }
 
     @Override
@@ -80,6 +153,7 @@ public class HouseDaoImpl implements HouseDao {
 
                 oldHouse.setHouseType(house.getHouseType());
                 oldHouse.setFurniture(house.isFurniture());
+                oldHouse.setRating(house.getRating());
                 oldHouse.setRoom(house.getRoom());
 
             }else {
@@ -92,21 +166,88 @@ public class HouseDaoImpl implements HouseDao {
             System.out.println(e.getMessage());
         }
 
-        System.out.println("updated to " + house.getHouse());
+        System.out.println("updated ");
     }
 
     @Override
     public List<House> getHousesByRegion(String region) {
-        return null;
+        EntityManager entityManager = entityManagerFactory.createEntityManager();
+
+        List<House> houses = new ArrayList<>();
+
+        try {
+            entityManager.getTransaction().begin();
+
+            List addresses = entityManager.createQuery("select a from Address  a where a.region =:par")
+                    .setParameter("par", region)
+                    .getResultList();
+
+            for (Object address : addresses) {
+//                houses.addAll(addresses.get)
+            }//////////////////////////////////////////////////
+
+            entityManager.getTransaction().commit();
+        } catch (Exception e) {
+            if (entityManager.getTransaction().isActive()) entityManager.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        }
+
+        return houses;
     }
 
     @Override
     public List<House> getHousesByAgency(Long agencyId) {
-        return null;
+        EntityManager entityManager = entityManagerFactory.createEntityManager();
+
+        List<House> houses = new ArrayList<>();
+
+        try {
+            entityManager.getTransaction().begin();
+
+            List<RentInfo> rentInfos = entityManager.createQuery("select r from RentInfo r where r.agency.id = :par", RentInfo.class)
+                    .setParameter("par", agencyId)
+                    .getResultList();
+
+            for (RentInfo rentInfo : rentInfos) {
+                houses.add(rentInfo.getHouse());
+            }
+
+            entityManager.getTransaction().commit();
+        } catch (Exception e) {
+            if (entityManager.getTransaction().isActive()) entityManager.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        } finally {
+            entityManager.close();
+        }
+
+        return houses;
     }
 
     @Override
     public List<House> getHousesByOwner(Long ownerId) {
-        return null;
+        EntityManager entityManager = entityManagerFactory.createEntityManager();
+
+        List<House> houses = new ArrayList<>();
+
+        try {
+            entityManager.getTransaction().begin();
+
+            List<RentInfo> rentInfos = entityManager.createQuery("select r from RentInfo r where r.owner.id = :par", RentInfo.class)
+                    .setParameter("par", ownerId)
+                    .getResultList();
+
+            for (RentInfo rentInfo : rentInfos) {
+                houses.add(rentInfo.getHouse());
+            }
+
+            entityManager.getTransaction().commit();
+        } catch (Exception e) {
+            if (entityManager.getTransaction().isActive()) entityManager.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        } finally {
+            entityManager.close();
+        }
+
+        return houses;
     }
 }
